+++ ex03/Bureaucrat.hpp
@@ -13,9 +13,9 @@ class Bureaucrat {
 
 private:
 	const std::string name;
-	int grade;
-	static const int highestGrade;
-	static const int lowestGrade;
+	short grade;
+	static const short highestGrade;
+	static const short lowestGrade;
 
 public:
 	class GradeTooHighException : public std::exception {
@@ -42,7 +42,7 @@ public:
 
 	Bureaucrat();
 
-	Bureaucrat(const std::string name, int grade);
+	Bureaucrat(const std::string name, short grade);
 
 	Bureaucrat(const Bureaucrat& other);
 
@@ -58,11 +58,11 @@ public:
 
 	const std::string getName() const;
 
-	int getGrade() const;
+	short getGrade() const;
 
-	static int getHighestGrade();
+	static short getHighestGrade();
 
-	static int getLowestGrade();
+	static short getLowestGrade();
 
 	void executeForm(AForm const& form);
